🚨 move to .svelte.ts
